Move NewGame into a different directory
