+++ src/components/CurrentLocation/CurrentLocation.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-// import { getStartingLocation } from "../../assets/api.js";
-import { mockGetStartingLocation } from "../../assets/mockAPI.js";
+import { getStartingLocation } from "../../assets/api.js";
+// import { mockGetStartingLocation } from "../../assets/mockAPI.js";
 import LocationInfo from "../LocationInfo/LocationInfo";
 
 function GetCurrentLocation({ token, info }: any) {
@@ -15,7 +15,7 @@ function GetCurrentLocation({ token, info }: any) {
             {!isRendered && (
                 <button
                     onClick={async () => {
-                        const json = await mockGetStartingLocation(
+                        const json = await getStartingLocation(
                             token,
                             nav_info.systemSymbol,
                             nav_info.waypointSymbol
